+++ src/node/ext/ldap/__init__.py
@@ -13,8 +13,7 @@ from node.ext.ldap.base import (
     LDAPConnector,
     testLDAPConnectivity,
 )
-# XXX: later
-#from node.ext.ldap.schema import LDAPSchemaInfo
+from node.ext.ldap.schema import LDAPSchemaInfo
 from node.ext.ldap.session import LDAPSession
 from node.ext.ldap._node import (
     LDAPNodeAttributes,